Major code refactoring (#5)
This is a major rework of the whole repository, and adds an integrated viewer (called cosma), which has been developed by me. The viewer is optional, and requires SDL2, imgui, stb and OpenGL to work properly.

The structure of the project is now modular and has three components: core, cosma, main.

The CLI11 is now also integrated and allows for project scaffolding. Currently missing is the update to the documentation. You should not use any commit before this one!